+++ tests/test_utils.py
@@ -105,3 +105,73 @@ def s3_bucket():
         ut.S3_CLIENT.create_bucket(Bucket=bucket_name, CreateBucketConfiguration=location)
 
         yield bucket_name
+
+
+def test_get_frame_id():
+    info = {
+        'metadata': {
+            '': {
+                'frame_id': '11113',
+            },
+        },
+    }
+    with patch('osgeo.gdal.Info', return_value=info) as mock_info:
+        assert ut.get_frame_id('foo.tiff') == 11113
+        mock_info.assert_called_once_with('foo.tiff', format='json')
+
+    info = {
+        'metadata': {
+            '': {
+                'frame_id': '321',
+            },
+        },
+    }
+    with patch('osgeo.gdal.Info', return_value=info) as mock_info:
+        assert ut.get_frame_id('bar.tiff') == 321
+        mock_info.assert_called_once_with('bar.tiff', format='json')
+
+
+def test_get_west_most_point():
+    info = {
+        'wgs84Extent': {
+            'type': 'Polygon',
+            'coordinates': [
+                [
+                    [-122.201872, 41.1562811],
+                    [-122.201872, 39.2490859],
+                    [-118.7684233, 39.2490859],
+                    [-118.7684233, 41.1562811],
+                    [-122.201872, 41.1562811],
+                ]
+            ],
+        },
+    }
+    with patch('osgeo.gdal.Info', return_value=info) as mock_info:
+        assert ut.get_west_most_point('foo.tiff') == -122.201872
+        mock_info.assert_called_once_with('foo.tiff', format='json')
+
+    info = {
+        'wgs84Extent': {
+            'type': 'Polygon',
+            'coordinates': [
+                [
+                    [176.4143837, 53.0448494],
+                    [176.4418596, 50.8863044],
+                    [-179.4626094, 50.8340006],
+                    [-179.2889357, 52.9883554],
+                    [176.4143837, 53.0448494],
+                ]
+            ],
+        },
+    }
+    with patch('osgeo.gdal.Info', return_value=info) as mock_info:
+        assert ut.get_west_most_point('bar.tiff') == -179.4626094
+        mock_info.assert_called_once_with('bar.tiff', format='json')
+
+
+def test_get_common_direction():
+    assert ut.get_common_direction({1, 2, 3}) == 'ASCENDING'
+    assert ut.get_common_direction({1, 873, 46986}) == 'ASCENDING'
+    assert ut.get_common_direction({68, 69, 70}) == 'DESCENDING'
+    with pytest.raises(ValueError):
+        ut.get_common_direction({67, 68})